+++ src/components/com_kunena/template/crypsisb4/layouts/user/item/status/default.php
@@ -14,7 +14,7 @@ defined('_JEXEC') or die;
 use Joomla\CMS\Language\Text;
 
 $user        = isset($this->user) ? $this->user : KunenaUserHelper::getMyself();
-$status      = $user->status;
+$status      = $user->getStatus();
 $status_text = $user->getStatusText();
 $link        = $user->getURL();